Changeing wrapper to nwrapper
