+++ nwrapper/ngraph/node.cpp
