+++ nwrapper/ngraph/ops/exp.cpp
@@ -8,25 +8,25 @@
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either logress or implied.
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // ----------------------------------------------------------------------------
 
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <string>
-#include "ngraph/ops/log.hpp"
+#include "ngraph/ops/exp.hpp"
 
 namespace py = pybind11;
 namespace ngraph {
 namespace op {
 
-PYBIND11_MODULE(Log, mod) {
+PYBIND11_MODULE(Exp, mod) {
 
-    py::module::import("wrapper.ngraph.ops.Op");
+    py::module::import("nwrapper.ngraph.ops.Op");
 
-    py::class_<Log, std::shared_ptr<Log>, UnaryElementwiseArithmetic> log(mod, "Log");
-    log.def(py::init<const std::shared_ptr<ngraph::Node>& >());
+    py::class_<Exp, std::shared_ptr<Exp>, UnaryElementwiseArithmetic> exp(mod, "Exp");
+    exp.def(py::init<const std::shared_ptr<ngraph::Node>& >());
 }
 
 }}  // ngraph
