+++ nwrapper/ngraph/ops/sum.cpp
@@ -15,19 +15,21 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <string>
-#include "ngraph/ops/multiply.hpp"
+#include "ngraph/ops/sum.hpp"
 
 namespace py = pybind11;
 namespace ngraph {
 namespace op {
 
-PYBIND11_MODULE(Multiply, mod) {
+PYBIND11_MODULE(Sum, mod) {
 
-    py::module::import("wrapper.ngraph.ops.Op");
+    py::module::import("nwrapper.ngraph.ops.Op");
 
-    py::class_<Multiply, std::shared_ptr<Multiply>, BinaryElementwiseArithmetic> multiply(mod, "Multiply");
-    multiply.def(py::init<const std::shared_ptr<ngraph::Node>&,
-                             const std::shared_ptr<ngraph::Node>& >());
+    using AxisSet = std::set<size_t>;
+
+    py::class_<Sum, std::shared_ptr<Sum>, RequiresTensorViewArgs> sum(mod, "Sum");
+    sum.def(py::init<const std::shared_ptr<ngraph::Node>&,
+                     const AxisSet& >());
 }
 
 }}  // ngraph
