+++ nwrapper/ngraph/util.cpp
