+++ setup.py
@@ -44,66 +44,66 @@ include_dirs = [# Path to pybind11 headers
                 get_pybind_include(),
                 get_pybind_include(user=True)
                ]
-ext_modules = [Extension('wrapper.ngraph.types.TraitedType',
-                        ['wrapper/ngraph/types/element_type.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Parameter',
-                        ['wrapper/ngraph/ops/parameter.cpp'], include_dirs),
-               Extension('wrapper.ngraph.types.TensorViewType',
-                        ['wrapper/ngraph/types/type.cpp'], include_dirs),
-               Extension('wrapper.ngraph.runtime.TensorView',
-                        ['wrapper/ngraph/runtime/tensor_view.cpp'], include_dirs),
-               Extension('wrapper.ngraph.Function',
-                        ['wrapper/ngraph/function.cpp'], include_dirs),
-               Extension('wrapper.ngraph.Node',
-                        ['wrapper/ngraph/node.cpp'], include_dirs),
-               Extension('wrapper.ngraph.runtime.Manager',
-                        ['wrapper/ngraph/runtime/manager.cpp'], include_dirs),
-               Extension('wrapper.ngraph.runtime.Backend',
-                        ['wrapper/ngraph/runtime/backend.cpp'], include_dirs),
-               Extension('wrapper.ngraph.runtime.ExternalFunction',
-                        ['wrapper/ngraph/runtime/external_function.cpp'], include_dirs),
-               Extension('wrapper.ngraph.runtime.clsNDArray',
-                        ['wrapper/ngraph/runtime/ndarray.cpp'], include_dirs),
-               Extension('wrapper.ngraph.runtime.CallFrame',
-                        ['wrapper/ngraph/runtime/call_frame.cpp'], include_dirs),
-               Extension('wrapper.ngraph.runtime.ParameterizedTensorView',
-                        ['wrapper/ngraph/runtime/parameterized_tensor_view.cpp'], include_dirs),
-               Extension('wrapper.ngraph.Util',
-                        ['wrapper/ngraph/util.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Op',
-                        ['wrapper/ngraph/ops/op.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Add',
-                        ['wrapper/ngraph/ops/add.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Dot',
-                        ['wrapper/ngraph/ops/dot.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Maximum',
-                        ['wrapper/ngraph/ops/maximum.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Divide',
-                        ['wrapper/ngraph/ops/divide.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Minimum',
-                        ['wrapper/ngraph/ops/minimum.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Exp',
-                        ['wrapper/ngraph/ops/exp.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Sum',
-                        ['wrapper/ngraph/ops/sum.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Greater',
-                        ['wrapper/ngraph/ops/greater.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Log',
-                        ['wrapper/ngraph/ops/log.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Convert',
-                        ['wrapper/ngraph/ops/convert.cpp'], include_dirs), 
-               Extension('wrapper.ngraph.ops.Subtract',
-                        ['wrapper/ngraph/ops/subtract.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Reshape',
-                        ['wrapper/ngraph/ops/reshape.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Broadcast',
-                        ['wrapper/ngraph/ops/broadcast.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.Reduce',
-                        ['wrapper/ngraph/ops/reduce.cpp'], include_dirs),
-               Extension('wrapper.ngraph.ops.ParameterizedConstant',
-                        ['wrapper/ngraph/ops/constant.cpp'], include_dirs),
-               Extension('wrapper.ngraph.runtime.Utils',
-                        ['wrapper/ngraph/runtime/utils.cpp'], include_dirs)]
+ext_modules = [Extension('nwrapper.ngraph.types.TraitedType',
+                        ['nwrapper/ngraph/types/element_type.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Parameter',
+                        ['nwrapper/ngraph/ops/parameter.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.types.TensorViewType',
+                        ['nwrapper/ngraph/types/type.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.runtime.TensorView',
+                        ['nwrapper/ngraph/runtime/tensor_view.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.Function',
+                        ['nwrapper/ngraph/function.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.Node',
+                        ['nwrapper/ngraph/node.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.runtime.Manager',
+                        ['nwrapper/ngraph/runtime/manager.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.runtime.Backend',
+                        ['nwrapper/ngraph/runtime/backend.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.runtime.ExternalFunction',
+                        ['nwrapper/ngraph/runtime/external_function.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.runtime.clsNDArray',
+                        ['nwrapper/ngraph/runtime/ndarray.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.runtime.CallFrame',
+                        ['nwrapper/ngraph/runtime/call_frame.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.runtime.ParameterizedTensorView',
+                        ['nwrapper/ngraph/runtime/parameterized_tensor_view.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.Util',
+                        ['nwrapper/ngraph/util.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Op',
+                        ['nwrapper/ngraph/ops/op.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Add',
+                        ['nwrapper/ngraph/ops/add.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Dot',
+                        ['nwrapper/ngraph/ops/dot.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Maximum',
+                        ['nwrapper/ngraph/ops/maximum.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Divide',
+                        ['nwrapper/ngraph/ops/divide.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Minimum',
+                        ['nwrapper/ngraph/ops/minimum.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Exp',
+                        ['nwrapper/ngraph/ops/exp.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Sum',
+                        ['nwrapper/ngraph/ops/sum.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Greater',
+                        ['nwrapper/ngraph/ops/greater.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Log',
+                        ['nwrapper/ngraph/ops/log.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Convert',
+                        ['nwrapper/ngraph/ops/convert.cpp'], include_dirs), 
+               Extension('nwrapper.ngraph.ops.Subtract',
+                        ['nwrapper/ngraph/ops/subtract.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Reshape',
+                        ['nwrapper/ngraph/ops/reshape.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Broadcast',
+                        ['nwrapper/ngraph/ops/broadcast.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.Reduce',
+                        ['nwrapper/ngraph/ops/reduce.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.ops.ParameterizedConstant',
+                        ['nwrapper/ngraph/ops/constant.cpp'], include_dirs),
+               Extension('nwrapper.ngraph.runtime.Utils',
+                        ['nwrapper/ngraph/runtime/utils.cpp'], include_dirs)]
 
 # As of Python 3.6, CCompiler has a `has_flag` m ethod.
 # cf http://bugs.python.org/issue26689
